createOutlineSilhouetteTextureを新規作成し、アウトラインメッシュ生成の依存関係を整理
